+++ internal/checks/alerts_comparison.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/cloudflare/pint/internal/discovery"
 	"github.com/cloudflare/pint/internal/parser"
+	"github.com/cloudflare/pint/internal/parser/utils"
 
 	promParser "github.com/prometheus/prometheus/promql/parser"
 )
@@ -41,6 +42,18 @@ func (c ComparisonCheck) Check(_ context.Context, _ string, rule parser.Rule, _ 
 	}
 
 	expr := rule.Expr().Query
+
+	if n := utils.HasOuterBinaryExpr(expr); n != nil && n.Op == promParser.LOR {
+		if (hasComparision(n.LHS) == nil || hasComparision(n.RHS) == nil) && !isAbsent(n.LHS) && !isAbsent(n.RHS) {
+			problems = append(problems, Problem{
+				Fragment: rule.AlertingRule.Expr.Value.Value,
+				Lines:    rule.AlertingRule.Expr.Lines(),
+				Reporter: c.Reporter(),
+				Text:     "alert query uses 'or' operator with one side of the query that will always return a result, this alert will always fire",
+				Severity: rewriteSeverity(Warning, n.LHS, n.RHS),
+			})
+		}
+	}
 
 	if n := hasComparision(expr.Node); n != nil {
 		if n.ReturnBool && hasComparision(n.LHS) == nil && hasComparision(n.RHS) == nil {
@@ -89,8 +102,15 @@ func hasComparision(n promParser.Node) *promParser.BinaryExpr {
 	return nil
 }
 
+func isAbsent(node promParser.Node) bool {
+	if node, ok := node.(*promParser.Call); ok && (node.Func.Name == "absent") {
+		return true
+	}
+	return false
+}
+
 func hasAbsent(n *parser.PromQLNode) bool {
-	if node, ok := n.Node.(*promParser.Call); ok && (node.Func.Name == "absent") {
+	if isAbsent(n.Node) {
 		return true
 	}
 	for _, child := range n.Children {
@@ -99,4 +119,13 @@ func hasAbsent(n *parser.PromQLNode) bool {
 		}
 	}
 	return false
+}
+
+func rewriteSeverity(s Severity, nodes ...promParser.Node) Severity {
+	for _, node := range nodes {
+		if n, ok := node.(*promParser.Call); ok && n.Func.Name == "vector" {
+			return Bug
+		}
+	}
+	return s
 }